LayoutTests/fast/parser/block-nesting-cap-expected.txt

Test that the HTML parser does not allow the nesting depth of elements to exceed 512.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d511.parentNode === d509 is true
PASS d510.parentNode === d509 is true
PASS d511.previousSibling === d510 is true
PASS successfullyParsed is true

TEST COMPLETE
